other commands

+++ src/commands/vesna/index.ts
@@ -4,18 +4,12 @@ import { Conversation } from "@grammyjs/conversations";
 import { InlineKeyboard } from "grammy";
 import { getGeneratedImages } from "../../core/supabase/ai";
 
-async function playomConversation(conversation: Conversation<MyContext>, ctx: MyContext) {
+async function vesnaConversation(conversation: Conversation<MyContext>, ctx: MyContext) {
   const keyboard = new InlineKeyboard().text("Отмена", "cancel");
   await ctx.reply("Привет! Напиши промпт на английском для генерации изображения.", {
     reply_markup: keyboard,
   });
   const { message, callbackQuery } = await conversation.wait();
-
-  if (callbackQuery?.data === "cancel") {
-    await ctx.reply("Вы отменили генерацию изображения.");
-    return;
-  }
-
   const info = await getGeneratedImages(ctx.from?.id.toString() || "");
   const { count, limit } = info;
 
@@ -23,14 +17,17 @@ async function playomConversation(conversation: Conversation<MyContext>, ctx: My
     await ctx.reply("У вас не осталось использований. Пожалуйста, оплатите генерацию изображений.");
     return;
   }
+  if (callbackQuery?.data === "cancel") {
+    await ctx.reply("Вы отменили генерацию изображения.");
+    return;
+  }
 
   if (!message || !ctx.from?.id) return;
 
   const generatingMessage = await ctx.reply("Генерация изображения началась...");
-  const image = await generateImage(`TOK ${message.text}` || "", "playom", ctx.from?.id.toString());
+  const image = await generateImage(`TOK ${message.text}` || "", "vesna", ctx.from?.id.toString());
   await ctx.api.deleteMessage(ctx.chat?.id || "", generatingMessage.message_id);
   await ctx.replyWithPhoto(image);
-
   if (count < limit) {
     await ctx.reply(`У вас осталось ${limit - count} использований.`);
     return;
@@ -40,4 +37,4 @@ async function playomConversation(conversation: Conversation<MyContext>, ctx: My
   }
 }
 
-export { playomConversation };
+export { vesnaConversation };
